Refactor SRCaloPoint into its own file, add more info and add truth info.
