+++ sbnanaobj/StandardRecord/SRTrackCalo.cxx
@@ -10,18 +10,6 @@
 
 namespace caf
 {
-
-  SRCaloPoint::SRCaloPoint():
-    rr(std::numeric_limits<float>::signaling_NaN()),
-    dqdx(std::numeric_limits<float>::signaling_NaN()),
-    dedx(std::numeric_limits<float>::signaling_NaN()),
-    pitch(std::numeric_limits<float>::signaling_NaN()),
-    t(std::numeric_limits<float>::signaling_NaN()),
-    integral(std::numeric_limits<float>::signaling_NaN()),
-    sumadc(std::numeric_limits<float>::signaling_NaN()),
-    wire(-1)
-  {}
-
   SRTrackCalo::SRTrackCalo():
     nhit(-999),
     ke(std::numeric_limits<float>::signaling_NaN()),
