+++ sbnanaobj/StandardRecord/SRTrackCalo.h
@@ -7,32 +7,10 @@
 #include <vector>
 #include "sbnanaobj/StandardRecord/SRVector3D.h"
 
+#include "sbnanaobj/StandardRecord/SRCaloPoint.h"
+
 namespace caf
 {
-
-  class SRCaloPoint {
-    public:
-
-      SRCaloPoint();
-      virtual ~SRCaloPoint() {}
-
-      float rr; //!< Residual Range [cm]
-      float dqdx; //!< dE/dx [MeV/cm]
-      float dedx; //!< dQ/dx [ADC/cm] -- pre calibration and electron lifetime correction
-      float pitch; //!< Track pitch [cm]
-      float t; //!< Time of deposition [ticks]
-      // NOTE: flatten-caf seems to have trouble with nesting the point
-      // inside a SRVector3D -- so just expose x/y/z out here
-      float x; //!< X-Position of deposition [cm]
-      float y; //!< Y-Position of deposition [cm]
-      float z; //!< Z-Position of deposition [cm]
-      float integral; //!< Hit Charge Integral [ADC]
-      float sumadc; //!< Hit Charge SummedADC [ADC]
-      short wire; //!< Wire of Calo-Point
-      short tpc; //!< TPC of Calo-Point
-  };
-
-
   /// Calorimetry information
   class SRTrackCalo
     {
